clean python tools
