--- tools/bn254_py/__init__.py
@@ -1,15 +1,0 @@
-"""Gives users direct access to classes and functions."""
-from bn254_py import big
-from bn254_py import bls
-from bn254_py import curve
-from bn254_py import ecp
-from bn254_py import ecp2
-from bn254_py import pair
-from bn254_py.ecp import ECp
-from bn254_py.ecp import generator
-from bn254_py.ecp2 import ECp2
-from bn254_py.fp import Fp
-from bn254_py.fp12 import Fp12
-from bn254_py.fp12 import one
-from bn254_py.fp2 import Fp2
-from bn254_py.fp4 import Fp4
